🚚 move page to canvas component
